requirements.txt
flask
python-dotenv
scikit-learn
pandas
numpy
joblib
requests
datasets
gensim
sentence-transformers
torch